tests/requirements.txt
pytest
pytest-cov
pyfakefs
freezegun
pytest-flask
selenium
webdriver-manager
requests
pytest-httpserver
docker
